feature: working on HSD.

+++ src/main/java/org/solitaire/model/Board.java
@@ -7,6 +7,5 @@ public interface Board<R> {
 
     List<R> path();
 
-    double score();
-
+    int score();
 }
